requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.5.2
httpx==0.28.1
google-genai==1.33.0
qdrant-client==1.11.3
SQLAlchemy==2.0.35
psycopg[binary]==3.2.1
alembic==1.13.2
tree-sitter==0.23.2
tree-sitter-languages==1.10.2
orjson==3.10.7
python-multipart==0.0.9
pytest==8.3.3
pytest-asyncio==0.24.0
